# Finished and Tested Item CRUD.

+++ application/src/main/java/com/aburakkontas/manga_payment/application/handlers/DeleteItemQueryHandler.java
@@ -2,13 +2,14 @@ package com.aburakkontas.manga_payment.application.handlers;
 
 import com.aburakkontas.manga.common.payment.queries.DeleteItemQuery;
 import com.aburakkontas.manga.common.payment.queries.results.DeleteItemQueryResult;
+import com.aburakkontas.manga_payment.domain.exceptions.ExceptionWithErrorCode;
 import com.aburakkontas.manga_payment.domain.repositories.ItemRepository;
 import org.axonframework.queryhandling.QueryHandler;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 
-@Service
+@Component
 public class DeleteItemQueryHandler {
 
     private final ItemRepository itemRepository;
@@ -20,10 +21,15 @@ public class DeleteItemQueryHandler {
 
     @QueryHandler
     public DeleteItemQueryResult handle(DeleteItemQuery query) {
+        var isExists = itemRepository.existsById(query.getItemId());
+
+        if(!isExists) throw new ExceptionWithErrorCode("Item not found", 404);
+
         itemRepository.deleteById(query.getItemId());
 
         var queryResult = new DeleteItemQueryResult();
         queryResult.setDeleted(true);
+        queryResult.setItemId(query.getItemId());
 
         return queryResult;
     }
